Support replies and reactions
All text message events now include three extra fields - `reply_id`, `emoji` and
`packet_id`.

`reply_id` and `emoji` correspond to the same fields in the `Data`
protobuf and can be used to determine whether the message is a reply or
a reaction.

`packet_id` corresponds to the `id` field of the `MeshPacket` protobuf.

The three service actions that send messages (Send Text, Broadcast
Channel Message and Send Direct Message) now accept `reply_id` and
`emoji` parameters, which will be reflected in the transmitted packet.
This can be used to send replies and reactions.